requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
python-multipart
python-dotenv
openai
